Moving utils.hpp under tests/ and added DBG print in Init/Finalize
