+++ tests/AMSlib/ams_ete.cpp
@@ -16,7 +16,7 @@
 #include "AMS.h"
 #include "wf/debug.h"
 
-#include "utils.hpp"
+#include "../utils.hpp"
 
 void createUmpirePool(std::string parent_name, std::string pool_name)
 {
