+++ tests/AMSlib/ams_update_model.cpp
@@ -11,7 +11,7 @@
 #include <vector>
 #include <wf/resource_manager.hpp>
 
-#include "utils.hpp"
+#include "../utils.hpp"
 
 #define SIZE (32L)
