+++ tests/AMSlib/ams_uq_model.cpp
@@ -9,7 +9,7 @@
 #include <vector>
 #include <wf/resource_manager.hpp>
 
-#include "utils.hpp"
+#include "../utils.hpp"
 
 #define SIZE (32L * 1024L + 3L)
