+++ tests/AMSlib/cpu_packing_test.cpp
@@ -13,7 +13,7 @@
 #include <wf/data_handler.hpp>
 #include <wf/resource_manager.hpp>
 
-#include "utils.hpp"
+#include "../utils.hpp"
 
 #define SIZE (32 * 1024 + 1)
