+++ tests/AMSlib/torch_model.cpp
@@ -15,7 +15,7 @@
 #include <vector>
 #include <wf/resource_manager.hpp>
 
-#include "utils.hpp"
+#include "../utils.hpp"
 
 #define SIZE (32L * 1024L + 3L)
